OdbcException.cpp: fix assert on NOT NULL schema countraints violation
Reproducing the issue:
---------------------
assert is raised when column 'id' on table 'test' does not "Allow Nulls"
when the following statement is executed:
  INSERT INTO test ( [id] ) VALUES ( NULL )

Fix:
---
The fix raises an ODBC Exception "No available diagnostic" when SQLGetDiagRec() does not succeed.

Explanation:
-----------
SQLGetDiagRec() can fail if there are no, or no more, diagnostic messages available as
there can be zero or more diagnostic message available.

For more information on SQLGetDiagRec() check the following ressources:
- SQLGetDiagRec  : http://msdn.microsoft.com/en-us/library/windows/desktop/ms716256(v=vs.85).aspx
- SQLGetDiagField: http://msdn.microsoft.com/en-us/library/windows/desktop/ms710181(v=vs.85).aspx